Ensure todo.txt line is unchanged on no change
Previously the line was regenerated, so e.g. if the context was not at
the end of the line it was after the run. This mad the "git diff"'s very
noisy. I like to run git diff after the trello2misc import to see what
changed and I have many todo.txt entries that are not from trello.